reconfigured folder structure from src\wizard\helpers\index.ts to src\helpers\index.ts in order to accommodate addtional future helper files/functions
AT-6841
