+++ src/helpers/wizard/index.ts
@@ -1,0 +1,10 @@
+export const displayErrorInPanel = (
+  selectorId: string,
+  type: string
+): boolean => {
+  return (
+    document.querySelector(
+      "[id^='" + selectorId + "']." + type + " .error--text"
+    ) !== null
+  );
+};
