--- src/wizard/helpers/index.ts
@@ -1,10 +1,0 @@
-export const displayErrorInPanel = (
-  selectorId: string,
-  type: string
-): boolean => {
-  return (
-    document.querySelector(
-      "[id^='" + selectorId + "']." + type + " .error--text"
-    ) !== null
-  );
-};
